Stashing des changements non commités avant l'extraction de 'master'. Dépilez ou appliquez ces changements pour les restaurer.

+++ headers/Timer.h
@@ -9,7 +9,7 @@ namespace gee
 		Timer(const std::string& name);
 		const std::string& name() const;
 		void reset();
-		const long long ellapsedMs() const;
+		const float ellapsedMs() const;
 		const long long ellapsedSec() const;
 	private:
 		std::string name_;
